refactor: ImageConverter 책임을  DIaryS3Service로 이동
